pyproject.toml
[tool.poetry]
name = "comtreigx"
version = "0.1.0"
description = ""
authors = ["Louisa Müller <louisa.mlr96@gmail.com>", "Philipp C. Verpoort <philipp.verpoort@pik-potsdam.de>"]
readme = "README.md"
packages = [
    {include = "comtreigx", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyyaml = "^6.0.1"
pandas = "^2.1.4"
comtradeapicall = "^1.0.20"
jupyterlab = "^3.5.2"
ipykernel = "^6.29.0"
pycountry = "^23.12.11"
matplotlib = "^3.8.2"
plotly = "^5.20.0"
seaborn = "^0.13.2"
scipy = "^1.13.0"
si-prefix = "^1.2.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
jupyterlab = "3.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
comtreigx = 'comtreigx.cmdline:main'
